start getting nds tests integrated
