+++ loader/src/manifest_nds.hpp
@@ -18,24 +18,60 @@
 #include <string>
 
 #include "manifest.hpp"
+#include "async_manager.hpp"
+#include "buffer_batch.hpp"
 
 namespace nervana
 {
     class manifest_nds;
 }
 
-class nervana::manifest_nds : public nervana::manifest
+class nervana::manifest_nds : public nervana::async_manager_source<variable_buffer_array>,
+                              public nervana::manifest
 {
 public:
-    manifest_nds(const std::string& filename);
+    manifest_nds(const std::string& baseurl, const std::string& token, size_t collection_id, size_t block_size,
+                     size_t shard_count = 1, size_t shard_index = 0);
     ~manifest_nds() {}
+
+
+    variable_buffer_array* next() override;
+    void reset() override
+    {
+    }
+
+    size_t record_count() const override
+    {
+        return m_record_count;
+    }
+
+    size_t element_count() const override
+    {
+        return 2;
+    }
+
+
+
     std::string cache_id() override;
 
     // NDS manifests doesn't have versions since collections are immutable
     std::string version() override { return ""; }
     static bool is_likely_json(const std::string filename);
 
-    std::string baseurl;
-    std::string token;
-    int         collection_id;
+    void load_metadata();
+    void get(const std::string& url, std::stringstream& stream);
+    const std::string load_block_url(uint32_t block_num);
+    const std::string metadata_url();
+
+    std::string m_baseurl;
+    std::string m_token;
+    int         m_collection_id;
+    size_t      m_block_size;
+    const size_t             m_shard_count;
+    const size_t             m_shard_index;
+    size_t m_record_count;
+    size_t m_block_count;
+
+private:
+    static size_t write_data(void* ptr, size_t size, size_t nmemb, void* stream);
 };
